+++ service/matching/config/config.go
@@ -80,8 +80,7 @@ type (
 		// task gc configuration
 		MaxTimeBetweenTaskDeletes time.Duration
 
-		EnableServiceDiscoveryShutdown dynamicconfig.BoolPropertyFn
-		EnableTasklistOwnershipGuard   dynamicconfig.BoolPropertyFn
+		EnableTasklistOwnershipGuard dynamicconfig.BoolPropertyFn
 	}
 
 	ForwarderConfig struct {
@@ -157,8 +156,7 @@ func NewConfig(dc *dynamicconfig.Collection, hostName string) *Config {
 		EnableTasklistIsolation:         dc.GetBoolPropertyFilteredByDomain(dynamicconfig.EnableTasklistIsolation),
 		AllIsolationGroups:              mapIGs(dc.GetListProperty(dynamicconfig.AllIsolationGroups)()),
 		AsyncTaskDispatchTimeout:        dc.GetDurationPropertyFilteredByTaskListInfo(dynamicconfig.AsyncTaskDispatchTimeout),
-		EnableServiceDiscoveryShutdown:  dc.GetBoolProperty(dynamicconfig.MatchingEnableServiceDiscoveryShutdown),
-		EnableTasklistOwnershipGuard:    dc.GetBoolProperty(dynamicconfig.MatchingEnableServiceDiscoveryShutdown),
+		EnableTasklistOwnershipGuard:    dc.GetBoolProperty(dynamicconfig.MatchingEnableTasklistGuardAgainstOwnershipShardLoss),
 		HostName:                        hostName,
 		TaskDispatchRPS:                 100000.0,
 		TaskDispatchRPSTTL:              time.Minute,
